requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
anthropic>=0.40.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
google-api-python-client>=2.115.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
httpx>=0.26.0
apscheduler>=3.10.4
openai>=1.10.0
pydub>=0.25.1
python-dotenv>=1.0.0
cryptography>=41.0.0
dateparser>=1.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=24.1.0
tenacity>=8.2.0
pgvector>=0.2.4
